+++ TimePicker/TimePicker.js
@@ -7,6 +7,7 @@
  * @module sandstone/TimePicker
  * @exports TimePicker
  * @exports TimePickerBase
+ * @exports timeToLocaleString
  */
 
 import Pure from '@enact/ui/internal/Pure';
@@ -243,5 +244,21 @@ const TimePicker = Pure(
  * @public
  */
 
+ /**
+ * Converts a standard `Date` object into a locale-specific string.
+ *
+ * @type {Function}
+ * @memberof sandstone/TimePicker
+ * @param {Date} time `Date` to convert
+ * @returns {String?} Converted date or `null` if `date` is invalid
+ */
+const timeToLocaleString = (time) => {
+	if (!time) {
+		return null;
+	}
+
+	return dateTimeConfig.i18n().formatter.format(time);
+};
+
 export default TimePicker;
-export {TimePicker, TimePickerBase};
+export {TimePicker, TimePickerBase, timeToLocaleString};