+++ TimePicker/tests/TimePicker-specs.js
@@ -1,7 +1,12 @@
 import React from 'react';
 import {mount} from 'enzyme';
-import TimePicker from '../TimePicker';
+import ilib from 'ilib';
+
+import TimePicker, {timeToLocaleString} from '../TimePicker';
 import css from '../TimePicker.module.less';
+
+// Note: Tests pass 'locale' because there's no I18nDecorator to provide a value via context and
+// otherwise, nothing renders in the label.
 
 describe('TimePicker', () => {
 
@@ -11,7 +16,7 @@ describe('TimePicker', () => {
 		() => {
 			const handleChange = jest.fn();
 			const subject = mount(
-				<TimePicker onChange={handleChange} open title="Time" value={new Date(2000, 6, 15, 3, 30)} locale="en-US" />
+				<TimePicker onChange={handleChange} value={new Date(2000, 6, 15, 3, 30)} locale="en-US" />
 			);
 
 			const base = subject.find('DateComponentRangePicker').first();
@@ -26,7 +31,7 @@ describe('TimePicker', () => {
 
 	test('should accept a JavaScript Date for its value prop', () => {
 		const subject = mount(
-			<TimePicker open title="Date" value={new Date(2000, 0, 1, 12, 30)} locale="en-US" />
+			<TimePicker value={new Date(2000, 0, 1, 12, 30)} locale="en-US" />
 		);
 
 		const minutePicker = subject.find(`.${css.minutePicker}`).at(0);
@@ -40,7 +45,7 @@ describe('TimePicker', () => {
 	test('should set "hourAriaLabel" to hour picker', () => {
 		const label = 'custom hour aria-label';
 		const subject = mount(
-			<TimePicker hourAriaLabel={label} open title="Date" value={new Date(2000, 0, 1, 12, 30)} />
+			<TimePicker hourAriaLabel={label} value={new Date(2000, 0, 1, 12, 30)} />
 		);
 
 		const hourPicker = subject.find(`.${css.hourPicker}`).at(0);
@@ -54,7 +59,7 @@ describe('TimePicker', () => {
 	test('should set "meridiemAriaLabel" to meridiem picker', () => {
 		const label = 'custom meridiem aria-label';
 		const subject = mount(
-			<TimePicker meridiemAriaLabel={label} open title="Date" value={new Date(2000, 0, 1, 12, 30)} />
+			<TimePicker meridiemAriaLabel={label} value={new Date(2000, 0, 1, 12, 30)} />
 		);
 
 		const meridiemPicker = subject.find(`.${css.meridiemPicker}`).at(0);
@@ -68,7 +73,7 @@ describe('TimePicker', () => {
 	test('should set "minuteAriaLabel" to minute picker', () => {
 		const label = 'custom minute aria-label';
 		const subject = mount(
-			<TimePicker minuteAriaLabel={label} open title="Date" value={new Date(2000, 0, 1, 12, 30)} />
+			<TimePicker minuteAriaLabel={label} value={new Date(2000, 0, 1, 12, 30)} />
 		);
 
 		const minutePicker = subject.find(`.${css.minutePicker}`).at(0);
@@ -81,7 +86,7 @@ describe('TimePicker', () => {
 
 	test('should set "data-webos-voice-disabled" to hour picker when voice control is disabled', () => {
 		const subject = mount(
-			<TimePicker open title="Date" value={new Date(2000, 0, 1, 12, 30)} data-webos-voice-disabled />
+			<TimePicker value={new Date(2000, 0, 1, 12, 30)} data-webos-voice-disabled />
 		);
 
 		const hourPicker = subject.find(`.${css.hourPicker}`).at(0);
@@ -94,7 +99,7 @@ describe('TimePicker', () => {
 
 	test('should set "data-webos-voice-disabled" to merdiem picker when voice control is disabled', () => {
 		const subject = mount(
-			<TimePicker open title="Date" value={new Date(2000, 0, 1, 12, 30)} data-webos-voice-disabled />
+			<TimePicker value={new Date(2000, 0, 1, 12, 30)} data-webos-voice-disabled />
 		);
 
 		const meridiemPicker = subject.find(`.${css.meridiemPicker}`).at(0);
@@ -107,13 +112,38 @@ describe('TimePicker', () => {
 
 	test('should set "data-webos-voice-disabled" to minute picker when voice control is disabled', () => {
 		const subject = mount(
-			<TimePicker open title="Date" value={new Date(2000, 0, 1, 12, 30)} data-webos-voice-disabled />
+			<TimePicker value={new Date(2000, 0, 1, 12, 30)} data-webos-voice-disabled />
 		);
 
 		const minutePicker = subject.find(`.${css.minutePicker}`).at(0);
 
 		const expected = true;
 		const actual = minutePicker.prop('data-webos-voice-disabled');
+
+		expect(actual).toBe(expected);
+	});
+
+	test('should format a date the same as the label', () => {
+		const time = new Date(2000, 0, 1, 12, 30);
+		const subject = mount(
+			<TimePicker value={time} locale="en-US" />
+		);
+
+		const expected = subject.find('Heading').text();
+		const actual = timeToLocaleString(time);
+
+		expect(actual).toBe(expected);
+	});
+
+	test('should format a date the same as the label in another locale', () => {
+		ilib.setLocale('ar-SA');
+		const time = new Date(2000, 0, 1, 12, 30);
+		const subject = mount(
+			<TimePicker value={time} locale="ar-SA" />
+		);
+
+		const expected = subject.find('Heading').text();
+		const actual = timeToLocaleString(time);
 
 		expect(actual).toBe(expected);
 	});